backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Neo4j
neo4j==5.16.0

# ML and Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.12.0

# Deep Learning
torch>=2.8.0
torchvision>=0.23.0
torch-geometric>=2.4.0

# NLP and Embeddings
sentence-transformers>=5.0.0

# Google Cloud (optional)
google-cloud-aiplatform==1.38.1
google-generativeai>=0.3.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
loguru==0.7.2

# Data Download
kaggle==1.7.4.5

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1

